refactor(inspector): extract useAppBuilderState() from AppBuilderTab
Moves ~700 lines of orchestration out of `AppBuilderTab` into a reusable
hook so both the legacy tab and the new Playground can compose their own
JSX over the same engine:

- Tool fetching (`fetchTools`, `loadToolsUntilMatch`)
- Form-field generation on tool select
- Protocol detection (MCP Apps / OpenAI SDK)
- Tool execution + result-injection waiters
- All 5 inspector command handlers (selectTool, executeTool,
  renderToolResult, setAppContext, snapshotApp)
- Onboarding state machine integration
- Server-sync timeout
- `loadingState` discriminated union the consumer switches on
  (`ready` | `skeleton` | `sync-timed-out` | `no-server`)

Exposes `AppBuilderStateContext` + `AppBuilderStateProvider` so the
Playground can call the hook once at the tab level and share it with
the docked tools pane (calling twice would double-register inspector
command handlers).

`AppBuilderTab.tsx` becomes a thin presentational shell (~240 lines)
that switches on `state.loadingState.kind` and renders
`<PlaygroundLeft/>` + `<PlaygroundMain/>`. Behavior is unchanged.

`PlaygroundMain.tsx` gains `evalChatHandoff` + `onEvalChatHandoffConsumed`
props and an effect mirroring ChatTabV2's handoff logic so eval
"Continue in chat" lands here once the flag redirects #chat-v2 to
#playground.
